BACKEND/requirements.txt
aiosmtplib==3.0.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
black==25.1.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
Deprecated==1.3.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-mail==1.5.0
flake8==7.3.0
Flask==3.1.1
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
limits==5.6.0
Mako==1.3.10
mariadb==1.1.12
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.17.1
mypy_extensions==1.1.0
openai==2.6.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.4
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.4.1
python-dotenv==1.1.1
python-http-client==3.3.7
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
ruff==0.12.8
sendgrid==6.12.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
Werkzeug==3.1.3
wrapt==2.0.0
reportlab==4.0.0
